Reflected GUI change to internal board 2D array.
Added printing the board to the application console for easy debugging.
Added a method to set the player input to the board if:
- A valid row and column were given.
- The position on the board is an empty position.
then returned wether the board successfully set the input or not,
so to not switch players or update the GUI if the previous rules were not met.