Added board state for detecting a winner or a tie.
Added a state enum class to differentiate between game states.

Maintained an input count to detect a tie if the board is fully marked.

Added an update state method that will return a new state based on the
state of the board, it detect wether the game is still going or if the
current player won or if the board is fully marked and its a tie.

Removed tracking for generated file .pro.user

When a cell is clicked the board state is updated and checked to see if
its a final state, and declares the game state in a message box