Added a vector of cell references.
Created a vector of nested class cell that contains
a reference to a button and it's associated row and
column, the whole cell is then passed to the slot when
the button is clicked for easier management of buttons.

Used std::vector instead of QVector to use emplace_back
function for optimization.