Added dummy AI logic.
Added two hardcoded option for AI,
- AIoppenent, specifies whether AI is available as an oppenent or not.
- startWithAI, Makes the AI player X, so it plays first.

AI dummy logic: choose first non empty cell

Added a Signal to turn finished by human input, for AI to connect to it
if it is present as an oppenent.

Added a dummy board function to get the index chosen by the AI to play.

Refactored game state update to its own function to be called after
Human or AI input.